requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
langchain
langchain-community
chromadb
pypdf
python-dotenv
langchain-ollama
requests
beautifulsoup4
streamlit
